dbus v0.6.0版初次发布: # 重要功能优化 * 优化dbus打包,添加deployer打包工程； * 优化dbus部署,所有配置统一在conf/config.properties中配置； * 全量部分代码重构,优化mysql拉取逻辑,防止出现oom导致worker自动重启导致任务拉取失败； * 新增sinker模块,实现增量数据落hdfs；
# 功能新增
* 新增全量断点续传功能,针对分片数目较多的任务,可以续传未完成的分片任务；
* 自动部署ogg、canal：
  新增自动部署canal,自动添加ogg目标端replicat进程功能；
  新增ogg、canal自动部署、重新部署、自动漂移功能；
  新增集群服务器canal、ogg部署详情集中管理功能；
* 新增批量重启拓扑功能,方便批量升级dbus mysql-extractor,dispatcher-appender、splitter-puller；
* 新增数据线一键检测功能，包括拓扑、ogg、canal、主备延时状态自动获取，此功能目前只提供给心跳程序使用；
* 重新设计优化，拓扑停止、状态获取,不再依赖stormUI；
* 新增更新表结构功能，用于加表后version为null，源端和dbus表结构不一致时自动更新表结构；

+++ dbus-heartbeat/src/main/java/com/creditease/dbus/heartbeat/event/impl/EmitHeartBeatEvent.java
@@ -2,14 +2,14 @@
  * <<
  * DBus
  * ==
- * Copyright (C) 2016 - 2018 Bridata
+ * Copyright (C) 2016 - 2019 Bridata
  * ==
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -18,11 +18,8 @@
  * >>
  */
 
-package com.creditease.dbus.heartbeat.event.impl;
 
-import java.util.LinkedHashSet;
-import java.util.Set;
-import java.util.concurrent.CountDownLatch;
+package com.creditease.dbus.heartbeat.event.impl;
 
 import com.creditease.dbus.heartbeat.dao.IHeartBeatDao;
 import com.creditease.dbus.heartbeat.dao.impl.HeartBeatDaoImpl;
@@ -33,9 +30,11 @@ import com.creditease.dbus.heartbeat.vo.DsVo;
 import com.creditease.dbus.heartbeat.vo.MonitorNodeVo;
 import com.creditease.dbus.heartbeat.vo.PacketVo;
 
+import java.util.LinkedHashSet;
+import java.util.Set;
+import java.util.concurrent.CountDownLatch;
+
 /**
- *
- *
  * @author Liang.Ma
  * @version 1.0
  */
@@ -88,17 +87,23 @@ public class EmitHeartBeatEvent extends AbstractEvent {
 
             // boolean isMysql = StringUtils.contains(ds.getDriverClass(), "mysql");
             int cnt = dao.sendPacket(ds.getKey(), node.getDsName(), node.getSchema(), node.getTableName(), strPacket, ds.getType());
-            if (cnt ==1 && isFirst) {
+            if (cnt == 1 && isFirst) {
                 saveZk(path, strPacket);
             }
 
-            //emitCount++
-            long emitCount = ds.getEmitCount();
-            ds.setEmitCount(emitCount + 1);
+            /**
+             * 根据日志发现，如果插入心跳失败，在执行删除心跳老数据时，执行select获取最大id时抛出异常，
+             * 会导致下一次创建preparestatment时卡死，所以只在插入成功时才执行删除操作
+             */
+            if (cnt == 1) {
+                //emitCount++
+                long emitCount = ds.getEmitCount();
+                ds.setEmitCount(emitCount + 1);
 
-            //删除不需要的心跳数据, 第一次emit一定会试图删除旧的
-            if (emitCount % 1000 == 0) {
-                dao.deleteOldHeartBeat(ds.getKey(), ds.getType());
+                //删除不需要的心跳数据, 第一次emit一定会试图删除旧的
+                if (emitCount % 1000 == 0) {
+                    dao.deleteOldHeartBeat(ds.getKey(), ds.getType());
+                }
             }
 
             //LoggerFactory.getLogger().info("心跳数据发送{},数据包[{}].", (cnt == 1) ? "成功" : "失败", strPacket);
